Separate Kafka from Golang docker-compose file

+++ Django/.docker/start.dev.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 
-cp .env.example .env
+if [ ! -f ".env" ]; then
+  cp .env.example .env
+fi
 
 python -m pip install pipenv
 
@@ -11,3 +13,5 @@ python -m pipenv run python manage.py migrate
 python -m pipenv run python manage.py loaddata initial_data
 
 python -m pipenv run python manage.py runserver 0.0.0.0:8000
+
+# tail -f /dev/null
